… Sources/PackageCollections/Storage/Trie.swift
… Tests/PackageCollectionsTests/PackageCollectionsStorageTests.swift
… Tests/PackageCollectionsTests/TrieTests.swift