@yim-lee 2021-01-20
@swift-ci Please test CentOS 7 platform