@yim-lee 2021-01-21
Verified locally with Amazon Linux 2 that `PackageCollectionsTests` pass with https://github.com/apple/swift-package-manager/pull/3213/commits/47c89860644d946c80bc3b293f930e5e73ee1787.

We can't fallback to FTS3--while DDL statements work with some modifications, query results are different when run on different platforms.